Crop works perfect and good validation
