+++ .gitignore
@@ -5,3 +5,4 @@ main
 *.o
 .DS_store
 *.bmp
+bin/
