+++ include/generate_start_setup.h
@@ -1,0 +1,6 @@
+#ifndef GENERATE_START_SETUP_H
+#define GENERATE_START_SETUP_H
+
+Parameters GenerateParameters(const int argc, char** argv);
+
+#endif //GENERATE_START_SETUP_H
