+++ include/validation_data.h
@@ -3,7 +3,7 @@
 
 void ValidationInputData(const Parameters& parameters, const std::unordered_map<std::string, int>& isFilter);
 void ValidationAllSupportedFilters(const Parameters& parameters, const std::unordered_map<std::string, int>& isFilter);
-void ValidationFiltersFailedMessage(const Parameters& parameters, const std::unordered_map<std::string, int>& isFilter);
+// void ValidationFiltersFailedMessage(const Parameters& parameters);
 void ValidationAllParametersInFilter(const Parameters& parameters, const std::unordered_map<std::string, int>& isFilter);
 void ValidateInOutPaths(const Parameters& parameters);
